Move post login page to frontend (#108)
* Create a named function for getting logged-in user's email

* Create a new route and page to let users know they've logged in successfully

* Rename directory 'lib' to 'utils' to match graphql-server's convention

* Update imports to reflect the renaming of dir from 'lib' to 'utils'

* Rename route 'check-login' to 'get-user-email' to be more descriptive & delete 'post-login' route

* Improve SmileNavBar component's typing and responsiveness

This commit hides the user's email and 'Log out' button on mobile-sized
windows. This handles the case where the user's email is too long to fit
on the navbar in the post-login popup window that states 'You are logged
in'. Outside of this scenario, SMILE is a desktop-only application so
this change will not affect the user experience.