Remove dataloader package
We previously used dataloader to make a single query for samples via
the OGM. Otherwise, the `samples` and `samplesConnection` sub-queries
would be call it twice.

With the new approach, we use two slightly different queries to get
samples data and the count of samples. We could use dataloader to access
the oncotree cache, but this operation is very fast (< 0.05ms).